Do not add .git/safe/../../bin to $PATH
This change to the PATH environment variable was interacting with my
Cygwin setup very poorly. It was causing the ldrel binary to not
be found (even though 'which ldrel' works fine), which makes all of
my QNX 6.5.0 projects impossible to compile.

Since I wasn't using this feature anyway, it is an easy decision to
disable it.