zsh: Add fixname.zsh configuration script
The $NAME environment variable is very common in build scripts. Having
it prepopulated with the hostname (as is apparently the default in WSL2)
isn't particularly useful, and causes builds to break (since all targets
have approximately the same name, it is impossible to build the whole
system)

Also, the $USERNAME environment variable is commonly used in Linux shell
scripting and shell configuration. The WSL2 system seems to
automatically populate a similar variable called $LOGNAME, but it is
easier to simply enable the same name so that scripts can more easily be
made portable.