Improve display of previous commands in zsh
If the current buffer is empty (ie. simply hitting <ENTER> repeatedly),
then simply put a blank line in the command history. This can be used to
add more space between commands (visutally).

Commands which are executed have their prompt changed to a timestamp
which records when they were run.

The default (not blankline, or timestamp) is to print the complex and
dynamic default prompt. This should only occur on the line with the
active cursor (which currently is always pinned to the bottom of the
terminal window)