Remove 'g' prefix from GNU sed and which binaries
There was a period of time where the Homebrew installations of gnu-sed
and gnu-which (on MacOS) would add a 'g' prefix in front of their binary
names by default (ie. gsed & gwhich). This was kind of annoying and only
occurred on that one operating system. It seems that this was determined
to be a "bad idea", so they have been reverted to the original / standard
default names of sed and which.